fix(runtime): RunFor::Forever is not secretly bounded at 4.29B cycles
`Kernel::end_cycle` was a `u32::MAX` sentinel meaning "no cycle bound",
but the bound test is `cycles >= end_cycle` — so the sentinel *was* a
bound. A `RunFor::Forever` run terminated once the counter reached it,
returning `Ok(())` with every `stop`/`teardown` hook firing as though the
bound had been reached normally. `RunFor::Duration` left the same
sentinel in place, so a duration-bounded run stopped there too, whatever
duration it asked for.

Measured on a 4-core VM, a busy-spin realtime graph (an
`Activation::ALWAYS` poll source) turns ~20M cycles/second:

    spin cycles: 39,805,345 in 2.0007s => 19,895,960 cycles/sec
    RunFor::Forever RETURNED Ok(()) after 215.31s, 4294967295 polls

"Forever" was 3.6 minutes, and `RunFor::Duration(8 hours)` was also 3.6
minutes. That shape is exactly the latency-critical ingest path the
adapters recommend — aeron `Spin`, iceoryx2 `Spin`, FIX `AlwaysSpin` —
and a graph carrying live orders that stops on its own, silently and
successfully, is the worst failure mode available. Slower graphs reach it
too: ~72 minutes at 1M events/s, ~12 hours at 100k/s.

- `end_cycle` becomes `Option<u64>`; `None` never compares done, so an
  unbounded run has no cycle bound at all rather than a distant one.
  `end_time` keeps its `NanoTime::MAX` sentinel — that one is ~584 years
  out and genuinely unreachable.
- `Kernel::cycles` widens to `u64`, with a `cycles()` accessor.
- `Runner::run_dynamic`'s `between` hook takes `u64` for the same reason:
  it hands the caller a cycle number that used to wrap.

Three tests pin it: an unbounded run seeded past the old sentinel still
begins cycles, a duration-bounded run likewise, and `RunFor::Cycles(n)`
still stops at exactly `n` and still round-trips through `run_for()`.

+++ crates/wingfoil/src/interp.rs
@@ -3088,7 +3088,7 @@ impl Runner {
         mut between: F,
     ) -> Result<()>
     where
-        F: FnMut(&mut Extension<'_>, u32) -> Result<()>,
+        F: FnMut(&mut Extension<'_>, u64) -> Result<()>,
     {
         // Source/run-mode validation, identical to `run`.
         self.node_visits = 0;
@@ -3147,7 +3147,10 @@ impl Runner {
             // long dynamic run does not allocate and zero a graph-sized array on
             // every one of its cycles.
             let mut dirty: Vec<bool> = Vec::new();
-            let mut cycles: u32 = 0;
+            // `u64`, not `u32`: this is handed to the caller's `between` hook on
+            // every cycle, and a realtime dynamic graph outlives a `u32`
+            // (`Kernel::cycles` carries the same widening for the same reason).
+            let mut cycles: u64 = 0;
             loop {
                 let n = self.nodes.len();
                 if node_dirty.len() < n {
